+++ mercury/src/test/java/org/broadinstitute/gpinformatics/athena/presentation/orders/ProductOrderActionBeanTest.java
@@ -473,7 +473,7 @@ public class ProductOrderActionBeanTest {
             Mockito.when(mockQuoteService.getQuoteByAlphaId(Mockito.anyString()))
                     .thenThrow(new QuoteServerException("Quote Server should not be called for SAP Source"));
 
-            sapMockQuote = TestUtils.buildTestSapQuote(quoteId, BigDecimal.valueOf(30000), BigDecimal.valueOf(37387),
+            sapMockQuote = TestUtils.buildTestSapQuote(quoteId, 30000d, 37387d,
                     pdo,quoteItemsMatchOrderProducts, "GP01");
 
             Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString())).thenReturn(sapMockQuote);
@@ -1590,7 +1590,7 @@ public class ProductOrderActionBeanTest {
                 @Override
                 public SapQuote answer(InvocationOnMock invocationOnMock) throws Throwable {
                     return TestUtils.buildTestSapQuote(testQuoteIdentifier,
-                            extraOrderValues.add(getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote)), BigDecimal.valueOf(800000),
+                            extraOrderValues.add(getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote)).doubleValue(), 800000d,
                             testOrder, quoteItemsMatchOrderProducts, "GP01");
                 }
             });
@@ -1751,8 +1751,8 @@ public class ProductOrderActionBeanTest {
                     final BigDecimal totalOpenOrderValue =
                             getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote);
                     final SapQuote sapQuote = TestUtils.buildTestSapQuote(testOrder.getQuoteId(),
-                            totalOpenOrderValue,
-                            totalOpenOrderValue.add(new BigDecimal(testQuote.getQuoteFunding().getFundsRemaining())),
+                            totalOpenOrderValue.doubleValue(),
+                            totalOpenOrderValue.add(new BigDecimal(testQuote.getQuoteFunding().getFundsRemaining())).doubleValue(),
                             testOrder, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01");
                     return sapQuote;
                 }
@@ -2335,8 +2335,8 @@ public class ProductOrderActionBeanTest {
                     .thenAnswer(new Answer<SapQuote>() {
                         @Override
                         public SapQuote answer(InvocationOnMock invocationOnMock) throws Throwable {
-                            return TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues,
-                                    BigDecimal.valueOf(800000), testOrder, quoteItemsMatchOrderProducts, "GP01");
+                            return TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues.doubleValue(),
+                                    800000d, testOrder, quoteItemsMatchOrderProducts, "GP01");
                         }
                     });
         } else {
@@ -2361,7 +2361,7 @@ public class ProductOrderActionBeanTest {
                         testOrder), overrideCalculatedOrderValue.add(extraOrderValues).doubleValue());
             testQuote.setQuoteItems(quoteItems);
         } else {
-            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues, BigDecimal.valueOf(800000),
+            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues.doubleValue(), 800000d,
                     testOrder, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01"), 0,
                     testOrder), overrideCalculatedOrderValue.add(extraOrderValues).doubleValue());
         }
@@ -2370,7 +2370,7 @@ public class ProductOrderActionBeanTest {
             Assert.assertEquals(actionBean.estimateOutstandingOrders(testQuote, 0,
                         null), extraOrderValues.doubleValue());
         } else {
-            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues, BigDecimal.valueOf(800000),
+            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues.doubleValue(), 800000d,
                     testOrder, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01"), 0,
                     null),
                     extraOrderValues.doubleValue());
@@ -2390,7 +2390,7 @@ public class ProductOrderActionBeanTest {
             Assert.assertEquals(actionBean.estimateOutstandingOrders(testQuote, 0, testOrder),
                     getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote).add(extraOrderValues).doubleValue());
         } else {
-            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues, BigDecimal.valueOf(800000),
+            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues.doubleValue(), 800000d,
                     testOrder, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01"), 0, testOrder),
                     getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote).add(extraOrderValues).doubleValue());
         }
@@ -2401,7 +2401,7 @@ public class ProductOrderActionBeanTest {
             Assert.assertEquals(actionBean.estimateOutstandingOrders(testQuote, 0, testOrder),
                     getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote).add(extraOrderValues).doubleValue());
         } else {
-            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues, BigDecimal.valueOf(800000),
+            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues.doubleValue(), 800000d,
                     testOrder, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01"), 0, testOrder),
                     getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote).add(extraOrderValues).doubleValue());
         }
@@ -2412,7 +2412,7 @@ public class ProductOrderActionBeanTest {
             Assert.assertEquals(actionBean.estimateOutstandingOrders(testQuote, 0, testOrder),
                     getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote).add(extraOrderValues).doubleValue());
         } else {
-            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues, BigDecimal.valueOf(800000),
+            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues.doubleValue(), 800000d,
                     testOrder, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01"), 0, testOrder),
                     getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote).add(extraOrderValues).doubleValue());
         }
@@ -2426,7 +2426,7 @@ public class ProductOrderActionBeanTest {
             Assert.assertEquals(actionBean.estimateOutstandingOrders(testQuote, 0, testOrder),
                     getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote).add(extraOrderValues).doubleValue());
         } else {
-            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues, BigDecimal.valueOf(800000),
+            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues.doubleValue(), 800000d,
                     testOrder, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01"), 0, testOrder),
                     getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote).add(extraOrderValues).doubleValue());
         }
@@ -2437,7 +2437,7 @@ public class ProductOrderActionBeanTest {
             Assert.assertEquals(actionBean.estimateOutstandingOrders(testQuote, 0, testOrder),
                     getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote).add(extraOrderValues).doubleValue());
         } else {
-            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues, BigDecimal.valueOf(800000),
+            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues.doubleValue(), 800000d,
                     testOrder, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01"), 0, testOrder),
                     getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote).add(extraOrderValues).doubleValue());
         }
@@ -2446,7 +2446,7 @@ public class ProductOrderActionBeanTest {
             Assert.assertEquals(actionBean.estimateOutstandingOrders(testQuote, 0, testOrder),
                     getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote).add(extraOrderValues).doubleValue());
         } else {
-            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues, BigDecimal.valueOf(800000),
+            Assert.assertEquals(actionBean.estimateSapOutstandingOrders(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues.doubleValue(), 800000d,
                     testOrder, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01"), 0, testOrder),
                     getCalculatedOrderValue(priceList, testOrder, quoteSource, testQuote).add(extraOrderValues).doubleValue());
         }
@@ -2456,7 +2456,7 @@ public class ProductOrderActionBeanTest {
         if (quoteSource == ProductOrder.QuoteSourceType.QUOTE_SERVER) {
             actionBean.validateQuoteDetails(testQuote, 0);
         } else {
-            actionBean.validateSapQuoteDetails(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues, BigDecimal.valueOf(800000),
+            actionBean.validateSapQuoteDetails(TestUtils.buildTestSapQuote(testQuoteIdentifier, extraOrderValues.doubleValue(), 800000d,
                     testOrder, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01"), 0);
         }
 
@@ -2581,8 +2581,8 @@ public class ProductOrderActionBeanTest {
             Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString())).thenAnswer(new Answer<SapQuote>() {
                 @Override
                 public SapQuote answer(InvocationOnMock invocationOnMock) throws Throwable {
-                    return TestUtils.buildTestSapQuote(testOrder.getQuoteId(), BigDecimal.valueOf(100000),
-                            BigDecimal.valueOf(1000000),testOrder,quoteItemsMatchOrderProducts, "GP01");
+                    return TestUtils.buildTestSapQuote(testOrder.getQuoteId(), 100000d,
+                            1000000d, testOrder,quoteItemsMatchOrderProducts, "GP01");
                 }
             });
         } else {
@@ -3747,7 +3747,7 @@ public class ProductOrderActionBeanTest {
                     @Override
                     public SapQuote answer(InvocationOnMock invocationOnMock) throws Throwable {
                         return TestUtils.buildTestSapQuote("27000001",
-                                BigDecimal.valueOf(30000), BigDecimal.valueOf(37387), pdo,
+                                30000d, 37387d, pdo,
                                 TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, (quoteMatchSalesOrg) ? "GP01" : "GP03");
                     }
                 });
